initial version of rest_lipid_top.py it reads the lipids_43A1-S3.itp and creates the typeA and typeB for bonds, angles, dihedrals and etc. The name conversion between typeA and typeB is simple for 	bonds g--> h 	pairs no change 	angles g --> h 	dihedrals g--> h
Name for atoms need more consideration
we want to the typeB name create from mapping from typeA does not collide with the existed names in the force field.
We need to check all the atom types in ffG43A1-S3par.02.itp to find out if our mapping is valid.

Desgn:
	rest_lipid_top.py  : read toplopy file from lipid.top
	rest_force_field_top.py : read force field both bonded and noboned parameters; for gromos 43a1-s3 , the bonded and nonbonded parameters are in the same file. This is the only format that is supported right now